+++ packages/renderer/src/lib/container/ContainerActions.svelte
@@ -11,7 +11,7 @@ import {
   faTerminal,
   faTrash,
 } from '@fortawesome/free-solid-svg-icons';
-import { Dropdown } from '@podman-desktop/ui-svelte';
+import { DropdownMenu } from '@podman-desktop/ui-svelte';
 import { createEventDispatcher, onMount } from 'svelte';
 import { router } from 'tinro';
 
@@ -127,9 +127,9 @@ function deployToKubernetes(): void {
 
 // If dropdownMenu = true, we'll change style to the imported dropdownMenu style
 // otherwise, leave blank.
-let actionsStyle: typeof Dropdown | typeof FlatMenu;
+let actionsStyle: typeof DropdownMenu | typeof FlatMenu;
 if (dropdownMenu) {
-  actionsStyle = Dropdown;
+  actionsStyle = DropdownMenu;
 } else {
   actionsStyle = FlatMenu;
 }
