+++ packages/ui/package.json
@@ -56,9 +56,9 @@
       "types": "./dist/alert/ErrorMessage.svelte.d.ts",
       "svelte": "./dist/alert/ErrorMessage.svelte"
     },
-    "./Dropdown": {
-      "types": "./dist/dropdown/index.d.ts",
-      "svelte": "./dist/dropdown/index.js"
+    "./DropdownMenu": {
+      "types": "./dist/dropdownMenu/index.d.ts",
+      "svelte": "./dist/dropdownMenu/index.js"
     },
     "./icons": {
       "types": "./dist/icons/index.d.ts",
