+++ packages/ui/src/lib/dropdownMenu/index.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2023 Red Hat, Inc.
+ * Copyright (C) 2024 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,24 +15,9 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
+import DropdownMenu from './DropdownMenu.svelte';
+import DropDownMenuItem from './DropDownMenuItem.svelte';
 
-import '@testing-library/jest-dom/vitest';
-
-import { fireEvent, render, screen } from '@testing-library/svelte';
-import { expect, test, vi } from 'vitest';
-
-import DropdownMenu from './Dropdown.svelte';
-
-test('Expect the onBeforeToggle function to be called when the menu is clicked', async () => {
-  const onBeforeToggleMock = vi.fn();
-  render(DropdownMenu, {
-    onBeforeToggle: onBeforeToggleMock,
-  });
-
-  const toggleMenuButton = screen.getByRole('button', { name: 'kebab menu' });
-  expect(toggleMenuButton).not.toHaveClass('mr-2');
-  expect(toggleMenuButton).toBeInTheDocument();
-  await fireEvent.click(toggleMenuButton);
-
-  expect(onBeforeToggleMock).toHaveBeenCalled();
+export default Object.assign(DropdownMenu, {
+  Item: DropDownMenuItem,
 });
